pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "ml_pipeline_env"
version = "1.0.0"
description = "OpenEnv environment for debugging broken ML pipelines"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "torch>=2.0.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "openai>=1.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ml_pipeline_env*", "server*"]
